+++ app/build.gradle
@@ -22,9 +22,10 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
+    compile project(':waveprogress')
+    compile project(':hellocharts-library')
     compile 'com.jakewharton:butterknife:5.1.1'
     compile 'com.android.support:appcompat-v7:24.0.0'
     compile 'com.android.support:design:24.0.0'
-    compile project(':waveprogress')
-    compile project(':hellocharts-library')
+    compile 'com.android.support:support-v4:24.0.0'
 }
