+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,11 @@
 <resources>
-    <string name="app_name">Clock</string>
+    <string name="app_name">Tomato Go</string>
     <string name="action_settings">Settings</string>
     <string name="title_activity_main2">Main2Activity</string>
     <string name="title_activity_regisiter">regisiter</string>
     <string name="title_activity_add_task">AddTask</string>
     <string name="title_activity_edit_task">EditTask</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
